+++ lib/src/iset/iset.dart
@@ -158,6 +158,13 @@ class ISet<T> // ignore: must_be_immutable
         config: config ?? defaultConfig,
       );
 
+  /// Converts from JSon. Json serialization support for json_serializable with @JsonSerializable.
+  factory ISet.fromJson(dynamic json, T Function(Object?) fromJsonT) =>
+      ISet<T>((json as Iterable).map(fromJsonT));
+
+  /// Converts to JSon. Json serialization support for json_serializable with @JsonSerializable.
+  Object toJson(Object Function(T) toJsonT) => map(toJsonT).toList();
+
   /// See also: [ImmutableCollection], [ImmutableCollection.lockConfig],
   /// [ImmutableCollection.isConfigLocked],[flushFactor], [defaultConfig]
   static void resetAllConfigurations() {